[ads] size_t code health

+++ components/brave_ads/core/internal/account/confirmations/queue/queue_item/confirmation_queue_item_test_util.cc
@@ -37,7 +37,7 @@ void SaveConfirmationQueueItems(
 }
 
 void BuildAndSaveConfirmationQueueItems(const ConfirmationInfo& confirmation,
-                                        int count) {
+                                        size_t count) {
   const ConfirmationQueueItemList confirmation_queue_items =
       BuildConfirmationQueueItems(confirmation, count);
