+++ components/brave_ads/core/internal/ad_units/inline_content_ad/inline_content_ad_feature.h
@@ -14,11 +14,11 @@ namespace brave_ads {
 BASE_DECLARE_FEATURE(kInlineContentAdFeature);
 
 // Set to 0 to never cap.
-inline constexpr base::FeatureParam<int> kMaximumInlineContentAdsPerHour{
+inline constexpr base::FeatureParam<size_t> kMaximumInlineContentAdsPerHour{
     &kInlineContentAdFeature, "maximum_ads_per_hour", 6};
 
 // Set to 0 to never cap.
-inline constexpr base::FeatureParam<int> kMaximumInlineContentAdsPerDay{
+inline constexpr base::FeatureParam<size_t> kMaximumInlineContentAdsPerDay{
     &kInlineContentAdFeature, "maximum_ads_per_day", 20};
 
 }  // namespace brave_ads
